# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Marsool delivery app API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /restaurants/{id}:
    get:
      operationId: getRestaurant
      tags: [restaurants]
      summary: Get a restaurant by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Restaurant details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Restaurant"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /restaurants/{id}/menu:
    get:
      operationId: getRestaurantMenu
      tags: [restaurants]
      summary: Get menu items for a restaurant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Menu items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Restaurant:
      type: object
      required:
        - id
        - name
        - nameAr
        - category
        - categoryAr
        - rating
        - reviewCount
        - deliveryTime
        - deliveryFee
        - minOrder
        - image
        - tags
        - isOpen
      properties:
        id:
          type: string
        name:
          type: string
        nameAr:
          type: string
        category:
          type: string
        categoryAr:
          type: string
        rating:
          type: number
        reviewCount:
          type: integer
        deliveryTime:
          type: string
        deliveryFee:
          type: number
        minOrder:
          type: number
        image:
          type: string
        tags:
          type: array
          items:
            type: string
        isOpen:
          type: boolean
        discount:
          type: string
          nullable: true
    MenuItemOption:
      type: object
      required: [id, nameAr, extraPrice, isDefault, sortOrder]
      properties:
        id:
          type: string
        nameAr:
          type: string
        extraPrice:
          type: integer
        isDefault:
          type: boolean
        sortOrder:
          type: integer
    MenuItemOptionGroup:
      type: object
      required: [id, nameAr, sortOrder, options]
      properties:
        id:
          type: string
        nameAr:
          type: string
        sortOrder:
          type: integer
        options:
          type: array
          items:
            $ref: '#/components/schemas/MenuItemOption'
    MenuItem:
      type: object
      required:
        - id
        - restaurantId
        - name
        - nameAr
        - price
        - image
        - category
        - categoryAr
        - isPopular
      properties:
        id:
          type: string
        restaurantId:
          type: string
        name:
          type: string
        nameAr:
          type: string
        description:
          type: string
        descriptionAr:
          type: string
        price:
          type: number
        image:
          type: string
        category:
          type: string
        categoryAr:
          type: string
        subcategory:
          type: string
          nullable: true
        subcategoryAr:
          type: string
          nullable: true
        isPopular:
          type: boolean
        isAvailable:
          type: boolean
        isDeal:
          type: boolean
        dealPrice:
          type: number
          nullable: true
        optionGroups:
          type: array
          items:
            $ref: '#/components/schemas/MenuItemOptionGroup'
...